+++ vrp-cli/src/commands/generate.rs
@@ -6,7 +6,7 @@ use super::*;
 use std::io::BufReader;
 use vrp_cli::extensions::generate::generate_problem;
 use vrp_pragmatic::format::problem::{serialize_problem, Problem};
-use vrp_pragmatic::format::{CoordIndex, FormatError};
+use vrp_pragmatic::format::CoordIndex;
 use vrp_pragmatic::validation::ValidationContext;
 
 pub const FORMAT_ARG_NAME: &str = "FORMAT";
@@ -101,12 +101,7 @@ fn generate_problem_from_args(matches: &ArgMatches) -> Result<(Problem, String),
             let coord_index = CoordIndex::new(&problem);
             ValidationContext::new(&problem, None, &coord_index)
                 .validate()
-                .map_err(|errors| {
-                    format!(
-                        "generated problem has some validation errors:\n{}",
-                        FormatError::format_many(errors.as_slice(), "\n")
-                    )
-                })
+                .map_err(|errs| format!("generated problem has some validation errors:\n{errs}",))
                 .map(|_| (problem, input_format.to_owned()))
         },
     )