+++ vrp-pragmatic/src/format/problem/mod.rs
@@ -30,11 +30,11 @@ use self::problem_reader::{map_to_problem_with_approx, map_to_problem_with_matri
 /// Reads specific problem definition from various sources.
 pub trait PragmaticProblem {
     /// Reads problem defined in pragmatic format.
-    fn read_pragmatic(self) -> Result<CoreProblem, Vec<FormatError>>;
+    fn read_pragmatic(self) -> Result<CoreProblem, MultiFormatError>;
 }
 
 impl<R: Read> PragmaticProblem for (BufReader<R>, Vec<BufReader<R>>) {
-    fn read_pragmatic(self) -> Result<CoreProblem, Vec<FormatError>> {
+    fn read_pragmatic(self) -> Result<CoreProblem, MultiFormatError> {
         let problem = deserialize_problem(self.0)?;
 
         let mut matrices = vec![];
@@ -47,7 +47,7 @@ impl<R: Read> PragmaticProblem for (BufReader<R>, Vec<BufReader<R>>) {
 }
 
 impl<R: Read> PragmaticProblem for BufReader<R> {
-    fn read_pragmatic(self) -> Result<CoreProblem, Vec<FormatError>> {
+    fn read_pragmatic(self) -> Result<CoreProblem, MultiFormatError> {
         let problem = deserialize_problem(self)?;
 
         map_to_problem_with_approx(problem)
@@ -55,7 +55,7 @@ impl<R: Read> PragmaticProblem for BufReader<R> {
 }
 
 impl PragmaticProblem for (String, Vec<String>) {
-    fn read_pragmatic(self) -> Result<CoreProblem, Vec<FormatError>> {
+    fn read_pragmatic(self) -> Result<CoreProblem, MultiFormatError> {
         let problem = deserialize_problem(BufReader::new(self.0.as_bytes()))?;
 
         let mut matrices = vec![];
@@ -68,7 +68,7 @@ impl PragmaticProblem for (String, Vec<String>) {
 }
 
 impl PragmaticProblem for String {
-    fn read_pragmatic(self) -> Result<CoreProblem, Vec<FormatError>> {
+    fn read_pragmatic(self) -> Result<CoreProblem, MultiFormatError> {
         let problem = deserialize_problem(BufReader::new(self.as_bytes()))?;
 
         map_to_problem_with_approx(problem)
@@ -76,19 +76,19 @@ impl PragmaticProblem for String {
 }
 
 impl PragmaticProblem for (ApiProblem, Vec<Matrix>) {
-    fn read_pragmatic(self) -> Result<CoreProblem, Vec<FormatError>> {
+    fn read_pragmatic(self) -> Result<CoreProblem, MultiFormatError> {
         map_to_problem_with_matrices(self.0, self.1)
     }
 }
 
 impl PragmaticProblem for ApiProblem {
-    fn read_pragmatic(self) -> Result<CoreProblem, Vec<FormatError>> {
+    fn read_pragmatic(self) -> Result<CoreProblem, MultiFormatError> {
         map_to_problem_with_approx(self)
     }
 }
 
 impl PragmaticProblem for (ApiProblem, Option<Vec<Matrix>>) {
-    fn read_pragmatic(self) -> Result<CoreProblem, Vec<FormatError>> {
+    fn read_pragmatic(self) -> Result<CoreProblem, MultiFormatError> {
         if let Some(matrices) = self.1 {
             (self.0, matrices).read_pragmatic()
         } else {